requirements.txt
asgiref==3.2.10
Django==3.1.2
django-crispy-forms==1.10.0
django-taggit==1.3.0
djangorestframework==3.12.2
Pillow==8.1.0
pytz==2020.1
sqlparse==0.4.1
